refactor: Renamed habit tracker to practice tracker
