+++ planner/python/classes/constants/entries.py
@@ -38,7 +38,7 @@ from classes.constants.addl_arg_keys import AddlArgKeys as Key
 from classes.constants.strings import PlannerStrings as Strings
 
 from classes.page_entries.day_entry_0 import DayEntry
-from classes.page_entries.daily_habit_tracker import DailyPracticeTracker
+from classes.page_entries.daily_practice_tracker import DailyPracticeTracker
 from classes.page_entries.free_write_entry import FreeWriteEntry
 from classes.page_entries.free_write_prompt_entry import FreeWritePromptEntry
 from classes.page_entries.goal_entry import GoalEntry
